+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 Version 7.5-SNAPSHOT
 -------------
 
+ADDED:
+ 
+- added new factory class for generate pins from given dates and date patterns
+- added new test dependency time-machine in version 1.2
+
 Version 7.4
 -------------